backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.0.1
requests==2.32.5
sqlalchemy==2.0.45
alembic==1.18.1
psycopg2-binary==2.9.11
pydantic-settings==2.12.0
python-multipart==0.0.20

# AI-парсер финансовых отчётов
pymupdf==1.24.13
openai==1.60.0
tenacity==9.0.0